refactor: AppInputPassword 수정 #85
